refactor: api documentation&change deleteChannel method

+++ src/main/java/com/messenger/chatty/service/ChannelServiceImpl.java
@@ -78,9 +78,10 @@ public class ChannelServiceImpl implements ChannelService{
 
 
     @Override
-    public void deleteChannelInWorkspace(Long channelId) {
+    public void deleteChannelInWorkspace(Long workspaceId, Long channelId) {
         Channel channel = channelRepository.findById(channelId)
                 .orElseThrow(() -> new CustomNoSuchElementException("채널ID", channelId, "채널"));
+        if(!channel.getWorkspace().getId().equals(workspaceId)) throw new CustomNoSuchElementException("워크스페이스 내에 해당 채널이 없습니다.");
         channelRepository.delete(channel);
     }
 }
